% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mixFil}
\alias{mixFil}
\title{All data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 370 rows and 3 columns.
}
\usage{
data(mixFil)
}
\description{
All data
}
\keyword{datasets}
